requirements.txt
pytest-pylint==0.16.1
pytest-flake8==1.0.5
PyGithub~=1.52